Switched from floats to doubles. Big fucking difference

+++ src/game.h
@@ -6,7 +6,7 @@ void draw_game(struct GameSharedData *);
 void draw_stat_bar(struct GameSharedData *);
 void request_game(struct GameSharedData *);
 void request_pause(struct GameSharedData *);
-float VectorAngle(float, float);
+double VectorAngle(double, double);
 bool get_drag_data(struct movable_object_structure *Obj, double *vx, double *vy, double *Cx, double *S);
 double coefficient_multiplier(double v);
 void* main_iteration(ALLEGRO_THREAD *, void *);
